Commit badger transaction synchronously
If badger.Txn.Commit is called with a non-nil
callback it writes to the database in the
background. Any error during write will be
written to the callback but the local err
variable will have gone out-of-scope by then.

For now, just provide a nil callback and then
Commit will write to the database
synchronously and Triplestore can handle any
error during write.